# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @nullable */
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface UserProfile {
  id: number;
  userId: string;
  name: string;
  age: number;
  gender: string;
  heightCm: number;
  weightKg: number;
  fitnessGoal: string;
  experienceLevel: string;
  activityLevel: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpsertProfileBody {
  name: string;
  age: number;
  gender: string;
  heightCm: number;
  weightKg: number;
  fitnessGoal: string;
  experienceLevel: string;
  activityLevel: string;
}

export interface Exercise {
  id: number;
  name: string;
  muscleGroup: string;
  /** @nullable */
  secondaryMuscles?: string | null;
  equipment: string;
  instructions: string;
  difficulty: string;
  isFavorite: boolean;
}

export interface ProgressionEntry {
  date: string;
  maxWeight: number;
  totalVolume: number;
  sets: number;
}

export interface ExerciseProgression {
  exerciseId: number;
  exerciseName: string;
  /** @nullable */
  personalRecord?: number | null;
  /** @nullable */
  lastWeight?: number | null;
  history: ProgressionEntry[];
}

export interface Workout {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  muscleGroup?: string | null;
  date: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  caloriesBurned?: number | null;
  /** @nullable */
  notes?: string | null;
  setCount: number;
  exerciseCount: number;
  createdAt: string;
}

export interface WorkoutSetDetail {
  id: number;
  workoutId: number;
  exerciseId: number;
  exerciseName: string;
  muscleGroup: string;
  setNumber: number;
  reps: number;
  weightKg: number;
  /** @nullable */
  notes?: string | null;
}

export interface WorkoutDetail {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  muscleGroup?: string | null;
  date: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  caloriesBurned?: number | null;
  /** @nullable */
  notes?: string | null;
  sets: WorkoutSetDetail[];
  createdAt: string;
}

export interface WorkoutSet {
  id: number;
  workoutId: number;
  exerciseId: number;
  setNumber: number;
  reps: number;
  weightKg: number;
  /** @nullable */
  notes?: string | null;
}

export interface CreateWorkoutBody {
  name: string;
  /** @nullable */
  muscleGroup?: string | null;
  date: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  caloriesBurned?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface AddSetBody {
  exerciseId: number;
  setNumber: number;
  reps: number;
  weightKg: number;
  /** @nullable */
  notes?: string | null;
}

export interface BodyweightLog {
  id: number;
  userId: string;
  weightKg: number;
  date: string;
  createdAt: string;
}

export interface AddBodyweightBody {
  weightKg: number;
  date: string;
}

export interface Measurement {
  id: number;
  userId: string;
  date: string;
  /** @nullable */
  chestCm?: number | null;
  /** @nullable */
  waistCm?: number | null;
  /** @nullable */
  hipsCm?: number | null;
  /** @nullable */
  bicepCm?: number | null;
  /** @nullable */
  thighCm?: number | null;
  createdAt: string;
}

export interface AddMeasurementBody {
  date: string;
  /** @nullable */
  chestCm?: number | null;
  /** @nullable */
  waistCm?: number | null;
  /** @nullable */
  hipsCm?: number | null;
  /** @nullable */
  bicepCm?: number | null;
  /** @nullable */
  thighCm?: number | null;
}

export interface WaterEntry {
  id: number;
  amountMl: number;
  loggedAt: string;
}

export interface WaterLog {
  date: string;
  totalMl: number;
  goalMl: number;
  logs: WaterEntry[];
}

export interface LogWaterBody {
  amountMl: number;
  date: string;
}

export interface PersonalRecord {
  exerciseId: number;
  exerciseName: string;
  weightKg: number;
  date: string;
}

export interface WeeklyVolumeEntry {
  day: string;
  volume: number;
  workouts: number;
}

export interface DashboardStats {
  totalWorkouts: number;
  weeklyStreak: number;
  currentStreak: number;
  totalCaloriesBurned: number;
  recentWorkouts: Workout[];
  personalRecords: PersonalRecord[];
  weeklyVolume: WeeklyVolumeEntry[];
}

export interface MealFood {
  name: string;
  amount: string;
  calories: number;
}

export interface Meal {
  id: number;
  mealPlanId: number;
  mealType: string;
  name: string;
  calories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
  foods: MealFood[];
}

export interface MealPlan {
  id: number;
  userId: string;
  dailyCalories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
  meals: Meal[];
  generatedAt: string;
}

export interface FoodLogEntry {
  id: number;
  userId: string;
  date: string;
  foodName: string;
  calories: number;
  /** @nullable */
  proteinG?: number | null;
  /** @nullable */
  carbsG?: number | null;
  /** @nullable */
  fatG?: number | null;
  /** @nullable */
  servingSize?: string | null;
  mealType: string;
  loggedAt: string;
}

export interface AddFoodLogBody {
  date: string;
  foodName: string;
  calories: number;
  /** @nullable */
  proteinG?: number | null;
  /** @nullable */
  carbsG?: number | null;
  /** @nullable */
  fatG?: number | null;
  /** @nullable */
  servingSize?: string | null;
  mealType: string;
}

export interface Achievement {
  id: number;
  key: string;
  name: string;
  description: string;
  icon: string;
  category: string;
  /** @nullable */
  earnedAt?: string | null;
  progress: number;
  target: number;
}

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type GetExercisesParams = {
  muscleGroup?: string;
  search?: string;
};