# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FitTrack Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProfileBody"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /exercises:
    get:
      operationId: getExercises
      tags: [exercises]
      summary: Get exercise library
      parameters:
        - name: muscleGroup
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
  /exercises/{id}/progression:
    get:
      operationId: getExerciseProgression
      tags: [exercises]
      summary: Get progression data for an exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise progression data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseProgression"
...
  /exercises/{id}/favorite:
    post:
      operationId: toggleFavoriteExercise
      tags: [exercises]
      summary: Toggle favorite status of exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Favorite toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /progress/bodyweight:
    get:
      operationId: getBodyweightLogs
      tags: [progress]
      summary: Get body weight log
      responses:
        "200":
          description: Body weight entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BodyweightLog"
    post:
      operationId: addBodyweightLog
      tags: [progress]
      summary: Log body weight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBodyweightBody"
      responses:
        "201":
          description: Weight logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BodyweightLog"
  /progress/measurements:
    get:
      operationId: getMeasurements
      tags: [progress]
      summary: Get body measurements
      responses:
        "200":
          description: Measurements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Measurement"
    post:
      operationId: addMeasurement
      tags: [progress]
      summary: Add body measurement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMeasurementBody"
      responses:
        "201":
          description: Measurement added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Measurement"
  /progress/water:
    get:
      operationId: getWaterLogs
      tags: [progress]
      summary: Get water intake for today
      responses:
        "200":
          description: Water logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterLog"
    post:
      operationId: logWater
      tags: [progress]
      summary: Log water intake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogWaterBody"
      responses:
        "200":
          description: Water logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterLog"
...
  /diet/food-log:
    get:
      operationId: getFoodLog
      tags: [diet]
      summary: Get food log for today
      responses:
        "200":
          description: Food log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodLogEntry"
    post:
      operationId: addFoodLog
      tags: [diet]
      summary: Add food log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFoodLogBody"
      responses:
        "201":
          description: Food logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodLogEntry"
  /diet/food-log/{id}:
    delete:
      operationId: deleteFoodLog
      tags: [diet]
      summary: Delete food log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        age:
          type: integer
        gender:
          type: string
        heightCm:
          type: number
        weightKg:
          type: number
        fitnessGoal:
          type: string
        experienceLevel:
          type: string
        activityLevel:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - name
        - age
        - gender
        - heightCm
        - weightKg
        - fitnessGoal
        - experienceLevel
        - activityLevel
        - createdAt
        - updatedAt
    UpsertProfileBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        gender:
          type: string
        heightCm:
          type: number
        weightKg:
          type: number
        fitnessGoal:
          type: string
        experienceLevel:
          type: string
        activityLevel:
          type: string
      required:
        - name
        - age
        - gender
        - heightCm
        - weightKg
        - fitnessGoal
        - experienceLevel
        - activityLevel
    Exercise:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        muscleGroup:
          type: string
        secondaryMuscles:
          type: ["string", "null"]
        equipment:
          type: string
        instructions:
          type: string
        difficulty:
          type: string
        isFavorite:
          type: boolean
      required:
        - id
        - name
        - muscleGroup
        - equipment
        - instructions
        - difficulty
        - isFavorite
    ExerciseProgression:
      type: object
      properties:
        exerciseId:
          type: integer
        exerciseName:
          type: string
        personalRecord:
          type: ["number", "null"]
        lastWeight:
          type: ["number", "null"]
        history:
          type: array
          items:
            $ref: "#/components/schemas/ProgressionEntry"
      required:
        - exerciseId
        - exerciseName
        - history
    ProgressionEntry:
      type: object
      properties:
        date:
          type: string
        maxWeight:
          type: number
        totalVolume:
          type: number
        sets:
          type: integer
      required:
        - date
        - maxWeight
        - totalVolume
        - sets
...
    BodyweightLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        weightKg:
          type: number
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - weightKg
        - date
        - createdAt
    AddBodyweightBody:
      type: object
      properties:
        weightKg:
          type: number
        date:
          type: string
      required:
        - weightKg
        - date
    Measurement:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
        chestCm:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        hipsCm:
          type: ["number", "null"]
        bicepCm:
          type: ["number", "null"]
        thighCm:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - date
        - createdAt
    AddMeasurementBody:
      type: object
      properties:
        date:
          type: string
        chestCm:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        hipsCm:
          type: ["number", "null"]
        bicepCm:
          type: ["number", "null"]
        thighCm:
          type: ["number", "null"]
      required:
        - date
    WaterLog:
      type: object
      properties:
        date:
          type: string
        totalMl:
          type: integer
        goalMl:
          type: integer
        logs:
          type: array
          items:
            $ref: "#/components/schemas/WaterEntry"
      required:
        - date
        - totalMl
        - goalMl
        - logs
    WaterEntry:
      type: object
      properties:
        id:
          type: integer
        amountMl:
          type: integer
        loggedAt:
          type: string
      required:
        - id
        - amountMl
        - loggedAt
    LogWaterBody:
      type: object
      properties:
        amountMl:
          type: integer
        date:
          type: string
      required:
        - amountMl
        - date
...
    FoodLogEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
        foodName:
          type: string
        calories:
          type: integer
        proteinG:
          type: ["integer", "null"]
        carbsG:
          type: ["integer", "null"]
        fatG:
          type: ["integer", "null"]
        servingSize:
          type: ["string", "null"]
        mealType:
          type: string
        loggedAt:
          type: string
      required:
        - id
        - userId
        - date
        - foodName
        - calories
        - mealType
        - loggedAt
    AddFoodLogBody:
      type: object
      properties:
        date:
          type: string
        foodName:
          type: string
        calories:
          type: integer
        proteinG:
          type: ["integer", "null"]
        carbsG:
          type: ["integer", "null"]
        fatG:
          type: ["integer", "null"]
        servingSize:
          type: ["string", "null"]
        mealType:
          type: string
      required:
        - date
        - foodName
        - calories
        - mealType
...