# lib/api-zod/src/generated/types/addBodyweightBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddBodyweightBody {
  weightKg: number;
  date: string;
}